Add more private function exclusions to config.jnaerator
There is likely a better way to do this, namely recognising the
G_GNUC_INTERNAL attribute. Skip them explicitly until it is figured
out how to do this.

The BridJ warnings about functions the addresses of which are not
found we prevent by doing this are those that we get when trying
to use the GObject bindings.